dbsp: Pick a consistent stream for trace bounds.
Since commit "dbsp: Keep traces that are likely to become large in
storage.", the following SQL test has failed with assertion "At least
one trace bound must be set" in trace.rs:

```
mvn test -Dtest="StreamingTests#profileLateness"
```

It's because that test applies `Stream::integrate_trace_retain_keys` to
one stream (to change its retainment functions) and another operator
internally creates a trace for a different stream, and the two uses end
up operating on different `TraceBounds` objects.

The problem arises from the different versions of streams that can
be created.  A stream can be sharded, and then it can be spilled to
storage.  Code tends to treat these streams interchangeably, but they
are not for this purpose.

This commit fixes the problem by separating bounds from traces.  Traces
need to be attached to the particular stream that they apply to, but
their bounds can be attached to the canonical version of the stream,
which this commit defines as the un-sharded, un-spilled version.  This
means that code that looks for the bounds, in order to add a retainment
function, will find the same `TraceBounds` that the trace did, even if
they are called for different versions.  The new `Stream::trace_bounds`
function finds (or creates) these canonical bounds.

This allowed for multiple additional simplifications.  First, both
`integrate_trace_retain_keys` and `integrate_trace_retain_values`
contained internal calls to `shard()`.  This was because of a
(previously removed) comment in those functions that I hadn't
understood:

```
        // The following `shard` is important.  It makes sure that the
        // bound is applied to the sharded version of the stream, which is what
        // all operators that come with a retainment policy use today.
        // If this changes in the future, we may need two versions of the operator,
        // with and without sharding.
        let (trace, bounds) = self.shard().integrate_trace_inner();
```

In other words, this issue came up before and it was solved by choosing
the sharded version as canonical.  This commit drops the sharding in
this case because it's no longer necessary.

Second, it made sense before to have both in-memory and on-storage
versions of most of the trace functions, because it was important to
both shard and spill in `integrate_trace_retain_keys` and
`integrate_trace_retain_values`.  This is no longer necessary for
correctness, so this commit drops the in-memory versions and changes
callers that want on-storage versions to just call `spill()` explicitly
beforehand.

Signed-off-by: Ben Pfaff <blp@feldera.com>