dbsp_adapters: Improve error handling in input adapters.
The controller's handling of fatal errors that occurred in input adapters
was poor.  If an input adapter failed and exited, such that it no longer
responded to input reader commands, then the controller could end up
waiting forever for the response.  This commit defines how input adapters
are supposed to behave when a fatal error occurs, in a way that is simple
enough that it should be hard to get wrong, and makes the controller
tolerate these fatal errors.

Thanks to Ashwin Kirpalani for providing a test case.

Signed-off-by: Ben Pfaff <blp@feldera.com>